+++ DarklyTests/Categories/LDClient+Testable.m
@@ -1,0 +1,13 @@
+//
+//  LDClient+Testable.m
+//  DarklyTests
+//
+//  Created by Mark Pokorny on 10/19/17. +JMJ
+//  Copyright © 2017 LaunchDarkly. All rights reserved.
+//
+
+#import "LDClient+Testable.h"
+
+@implementation LDClient(Testable)
+@dynamic clientStarted;
+@end