+++ DarklyTests/DarklyXCTestCase.m
@@ -13,7 +13,6 @@
     [super setUp];
 
     [[NSUserDefaults standardUserDefaults] removeObjectForKey:kUserDictionaryStorageKey];
-    [[LDDataManager sharedManager] flushEventsDictionary];
 }
 
 - (void)tearDown {