+++ DarklyTests/LDPollingManagerTest.m
@@ -4,39 +4,45 @@
 
 #import "DarklyXCTestCase.h"
 #import "LDPollingManager.h"
-#import "LDDataManager.h"
-#import "LDRequestManager.h"
+#import "LDClientManager.h"
 #import "LDConfig.h"
 #import "LDClient.h"
 #import "OCMock.h"
 
+extern NSString *const kTestMobileKey;
+
 @interface LDPollingManagerTest : DarklyXCTestCase
-@property (nonatomic) id requestManagerMock;
+@property (nonatomic, strong) id mockLDClient;
+@property (nonatomic, strong) id mockLDClientManager;
 @end
 
 @implementation LDPollingManagerTest
-@synthesize dataManagerMock;
-@synthesize requestManagerMock;
 
 - (void)setUp {
     [super setUp];
     
-    LDRequestManager *requestManager = [LDRequestManager sharedInstance];
-    requestManagerMock = OCMPartialMock(requestManager);
-    OCMStub([requestManagerMock performFeatureFlagRequest:[OCMArg isKindOfClass:[NSString class]]]);
+    id mockLDClient = OCMClassMock([LDClient class]);
+    OCMStub(ClassMethod([mockLDClient sharedInstance])).andReturn(mockLDClient);
+    self.mockLDClient = mockLDClient;
 
-    id requestManagerClassMock = OCMClassMock([LDRequestManager class]);
-    OCMStub(ClassMethod([requestManagerClassMock sharedInstance])).andReturn(requestManager);
+    id mockLDClientManager = OCMClassMock([LDClientManager class]);
+    OCMStub(ClassMethod([mockLDClientManager sharedInstance])).andReturn(mockLDClientManager);
+    self.mockLDClientManager = mockLDClientManager;
  }
 
 - (void)tearDown {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
-    [dataManagerMock stopMocking];
-    [requestManagerMock stopMocking];
+    [self.mockLDClient stopMocking];
+    [self.mockLDClientManager stopMocking];
     [super tearDown];
 }
 
 - (void)testEventPollingStates {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+
     // create the expectation with a nice descriptive message
     LDPollingManager *dnu =  [LDPollingManager sharedInstance];
     dnu.eventPollingIntervalMillis = 5000; // for the purposes of the unit tests set it to 5 secs.
@@ -63,34 +69,196 @@
 }
 
 - (void)testPollingInterval {
-    NSString *testMobileKey = @"testMobileKey";
-    LDConfig *config = [[LDConfig alloc] initWithMobileKey:testMobileKey];
-    LDClient *client = [LDClient sharedInstance];
-    LDPollingManager *pollingManager =  [LDPollingManager sharedInstance];
-    
-    [client start:config withUserBuilder:nil];
-    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, kDefaultFlushInterval*kMillisInSecs);
-    XCTAssertEqual([pollingManager configPollingState], POLL_STOPPED);
-    [client stopClient];
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
 
-    config.streaming = NO;
-    [client start:config withUserBuilder:nil];
-    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, kDefaultPollingInterval*kMillisInSecs);
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+    LDPollingManager *pollingManager =  [LDPollingManager sharedInstance];
+
+    [pollingManager startConfigPolling];
     XCTAssertEqual(pollingManager.configPollingIntervalMillis, kDefaultPollingInterval*kMillisInSecs);
     XCTAssertEqual([pollingManager configPollingState], POLL_RUNNING);
-    [client stopClient];
+    [pollingManager stopConfigPolling];
 
-    config.pollingInterval = [NSNumber numberWithInt:50];
-    [client start:config withUserBuilder:nil];
-    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, kMinimumPollingInterval*kMillisInSecs);
+    [pollingManager startEventPolling];
+    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, kDefaultFlushInterval*kMillisInSecs);
+    XCTAssertEqual([pollingManager eventPollingState], POLL_RUNNING);
+    [pollingManager stopEventPolling];
+
+    config.pollingInterval = [NSNumber numberWithInt:kMinimumPollingInterval - 1];
+    [pollingManager startConfigPolling];
     XCTAssertEqual(pollingManager.configPollingIntervalMillis, kMinimumPollingInterval*kMillisInSecs);
-    [client stopClient];
+    [pollingManager stopConfigPolling];
+
+    config.pollingInterval = [NSNumber numberWithInt:kMinimumPollingInterval + 1];
+    [pollingManager startConfigPolling];
+    XCTAssertEqual(pollingManager.configPollingIntervalMillis, [config.pollingInterval intValue]*kMillisInSecs);
+    [pollingManager stopConfigPolling];
 
     config.flushInterval = [NSNumber numberWithInt:50];
-    [client start:config withUserBuilder:nil];
-    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, 50*kMillisInSecs);
-    XCTAssertEqual(pollingManager.configPollingIntervalMillis, kMinimumPollingInterval*kMillisInSecs);
-    [client stopClient];
+    [pollingManager startEventPolling];
+    XCTAssertEqual(pollingManager.eventPollingIntervalMillis, [config.flushInterval intValue]*kMillisInSecs);
+    [pollingManager stopEventPolling];
+}
+
+- (void)testConfigPollingOnline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopConfigPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+
+    [pollingManager startConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_RUNNING);
+
+    [pollingManager stopConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+}
+
+- (void)testConfigPollingOffline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(NO);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopConfigPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+
+    [pollingManager startConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+}
+
+- (void)testResumeConfigPollingOnline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopConfigPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+
+    [pollingManager startConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_RUNNING);
+
+    [pollingManager suspendConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_SUSPENDED);
+
+    [pollingManager resumeConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_RUNNING);
+
+    [pollingManager stopConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+}
+
+- (void)testResumeConfigPollingOffline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    [[[self.mockLDClientManager expect] andReturnValue:OCMOCK_VALUE(YES)] isOnline];
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopConfigPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+
+    [pollingManager startConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_RUNNING);
+
+    [pollingManager suspendConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_SUSPENDED);
+
+    [[[self.mockLDClientManager expect] andReturnValue:OCMOCK_VALUE(NO)] isOnline];
+    XCTAssertTrue([LDClientManager sharedInstance].isOnline == NO);
+
+    [pollingManager resumeConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_SUSPENDED);
+
+    [pollingManager stopConfigPolling];
+    XCTAssertTrue(pollingManager.configPollingState == POLL_STOPPED);
+}
+
+- (void)testEventPollingOnline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopEventPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+
+    [pollingManager startEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_RUNNING);
+
+    [pollingManager stopEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+}
+
+- (void)testEventPollingOffline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(NO);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopEventPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+
+    [pollingManager startEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+}
+
+- (void)testResumeEventPollingOnline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    OCMStub([self.mockLDClientManager isOnline]).andReturn(YES);
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopEventPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+
+    [pollingManager startEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_RUNNING);
+
+    [pollingManager suspendEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_SUSPENDED);
+
+    [pollingManager resumeEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_RUNNING);
+
+    [pollingManager stopEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+}
+
+- (void)testResumeEventPollingOffline {
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:kTestMobileKey];
+    OCMStub([self.mockLDClient ldConfig]).andReturn(config);
+
+    [[[self.mockLDClientManager expect] andReturnValue:OCMOCK_VALUE(YES)] isOnline];
+
+    LDPollingManager *pollingManager = [LDPollingManager sharedInstance];
+    [pollingManager stopEventPolling]; //Do not assume anything about polling manager state
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
+
+    [pollingManager startEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_RUNNING);
+
+    [pollingManager suspendEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_SUSPENDED);
+
+    [[[self.mockLDClientManager expect] andReturnValue:OCMOCK_VALUE(NO)] isOnline];
+    XCTAssertTrue([LDClientManager sharedInstance].isOnline == NO);
+
+    [pollingManager resumeEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_SUSPENDED);
+
+    [pollingManager stopEventPolling];
+    XCTAssertTrue(pollingManager.eventPollingState == POLL_STOPPED);
 }
 
 @end